refactor: extract safe db fns
