+++ src/workers/classify.spec.ts
@@ -1,11 +1,11 @@
-import { Connection, createConnection } from 'typeorm';
-import { readFile, readFileSync } from 'fs-extra';
+import { readFileSync } from 'fs-extra';
 import { join } from 'path';
 import classify from './classify';
 import Message from '../entities/message';
 import { encode } from '../lib/util';
 import { parseKeywordLists, List } from '../lib/classify';
 import FAKE_RECEIVED_HEADERS from '../../fixtures/fake_received_headers.json';
+import ensureSafeDb from '../spec/helpers/ensure_safe_db';
 
 const messageFixtures = [
   // Empty; should be unclassifiable
@@ -56,26 +56,7 @@ function createMessages(): Promise<Message[]> {
 }
 
 describe('message db tests', () => {
-  let conn: Connection;
-  beforeAll(async () => {
-    conn = await createConnection();
-  });
-  afterAll(async () => {
-    await conn.close();
-  });
-
-  afterEach(async () => {
-    // HACK: This is a really jank way to ensure that VSCode's Jest runner doesn't nuke the current
-    // DB when you're running a dev server alongside your editor
-    const ormconfig = JSON.parse(
-      (await readFile('./ormconfig.json')).toString(),
-    );
-    const { database }: { database: string } = ormconfig;
-    if (!database.endsWith('_test')) {
-      throw new Error(`Cowardly refusing to clear non-test DB ${database}`);
-    }
-    await Message.clear();
-  });
+  ensureSafeDb();
 
   describe('classify', () => {
     let lists: List[];
