+++ src/workers/persist.spec.ts
@@ -1,5 +1,3 @@
-import { Connection, createConnection } from 'typeorm';
-import { readFile } from 'fs-extra';
 import persist, {
   SparseMessage,
   castToSparse,
@@ -12,6 +10,7 @@ import persist, {
 import { Message as GmailMessage, GmailClient } from '../types';
 import Message from '../entities/message';
 import FAKE_RECEIVED_HEADERS from '../../fixtures/fake_received_headers.json';
+import ensureSafeDb from '../spec/helpers/ensure_safe_db';
 
 jest.mock('../lib/gmail/api');
 
@@ -31,26 +30,7 @@ describe('castToSparse', () => {
 });
 
 describe('message db tests', () => {
-  let conn: Connection;
-  beforeAll(async () => {
-    conn = await createConnection();
-  });
-  afterAll(async () => {
-    await conn.close();
-  });
-
-  afterEach(async () => {
-    // HACK: This is a really jank way to ensure that VSCode's Jest runner doesn't nuke the current
-    // DB when you're running a dev server alongside your editor
-    const ormconfig = JSON.parse(
-      (await readFile('./ormconfig.json')).toString(),
-    );
-    const { database }: { database: string } = ormconfig;
-    if (!database.endsWith('_test')) {
-      throw new Error(`Cowardly refusing to clear non-test DB ${database}`);
-    }
-    await Message.clear();
-  });
+  ensureSafeDb();
 
   async function insertMessagesWithIds(...ids: string[]): Promise<void> {
     await Promise.all(
